+++ R/install-cran.R
@@ -38,7 +38,7 @@ install_cran <- function(pkgs, repos = getOption("repos"), type = getOption("pkg
                   ...)
 }
 
-cran_remote <- function(pkg, repos, type, ...) {
+cran_remote <- function(pkg, repos = getOption("repos"), type = getOption("pkgType"), ...) {
 
   remote("cran",
     name = pkg,